Fixed request corruption bug.
* curl option `CURLOPT_POSTFIELDS` is replaced by `CURLOPT_COPYPOSTFIELDS` to prevent post request corruption.

According to the official documentation, using `CURLOPT_POSTFIELDS` is required the pointed data to be preserved until request completely finished, which is not guaranteed while using asynchronous requests.
Instead of it, `CURLOPT_COPYPOSTFIELDS` is used: original data are copied by the library, allowing the application to overwrite the original data after setting this option.