connection: remove handler from the connection once handler's on_close() method is called.
Handler's on_close() method is called with error in case of read/write failure.
In this case handler is removed from the connection and associated socket is closed.

Handler's on_close() method is called with no error after all data from the client is received.
In this case handler is removed from the connection and from this point it's user's responsibility
to keep handler alive (with associated connection).