tests/fixtures/server: fix server's cleanup phase
@pytest.fixture is used for server instead of @pytest.yield_fixture to explicitly
specify finalization function via `request.addfinalizer`.
With @pytest.yield_fixture it's unclear if cleanup phase (that follows yield)
will be called in case of exceptions in setup phase.

In `Server.terminate()` only synchronous methods are allowed, because it's
possible for test's io_loop to be stopped before fixture's cleanup phase.