thevoid/connection: close connection gracefully after sending the response
When server attempts to terminate request processing before the whole request
was received it should close the connection gracefully after sending the
response.

To do this server will read remaining request's body until either the whole
request is received or the client closes the connection.

If server has received the whole request it may close the connection in a usual
way (gracefully shutting down the socket).

If during graceful close the server fails on receiving remaining data from the
client it means the client received the response and closed the connection by
themself. In this case the server should not consider this fail as an error.